Rename: validation/path에서 validation/pathname으로 수정
